==> 2-preprocessText.ipynb <==
import re
import pandas as pd
from io import StringIO
import nltk
import numpy as np
import json

nltk.download('stopwords')
nltk.download('punkt')
nltk.download('wordnet')

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import PorterStemmer

lemma = nltk.wordnet.WordNetLemmatizer()

ps = PorterStemmer()

stop_words = set(stopwords.words('english'))

def extractDataFrame(fileName):
  with open(fileName) as file:
    lines = [re.sub(r'([^,])"(\s*[^\n])', r'\1/"\2', line) for line in file]
    df = pd.read_csv(StringIO(''.join(lines)), escapechar="/")
  return df

def splitLabels(labels):
  return [int(label) for label in labels.split(' ')]

def tokenise(caption):
    # Replace non-alphabetic characters with single whitespace
    caption = re.sub(r'[^a-zA-Z\s]', ' ', caption.lower())
    # Remove any whitespace that appears in sequence
    caption = re.sub(r"\s+", " ", caption)
    # Remove new leading and trailing whitespace
    caption = caption.strip()
    # Apply stemmer to each remaining word in sentence
    caption = lemma.lemmatize(caption)
    caption = ps.stem(caption)

    # Tokenize
    word_tokens = word_tokenize(caption)
    # Remove stop words
    filtered_sentence = [w for w in word_tokens if not w.lower() in stop_words]
    # Replace value with tokenised data
    return filtered_sentence

train = extractDataFrame('data/train.csv')

train['Tokens'] = train['Caption'].apply(tokenise)
train['Labels'] = train['Labels'].apply(splitLabels)

uniqueLabels = set()
for labels in train['Labels']:
  uniqueLabels.update(labels)

numUniqueLabels = len(uniqueLabels)
maxLabel = max(uniqueLabels)

print(uniqueLabels)
print('Unique labels:', numUniqueLabels)
print('Max label value:', maxLabel)

def labelMapping(uLabels):
  labelToIdMapping = {}
  idToLabelMapping = {}
  for idx, label in enumerate(uLabels):
    labelToIdMapping[label] = idx
    idToLabelMapping[idx] = label
  return labelToIdMapping, idToLabelMapping

labelToIdMapping, idToLabelMapping = labelMapping(uniqueLabels)

print(idToLabelMapping)

def mapLabelToId(labels):
  ids = []
  for label in labels:
    ids.append(labelToIdMapping[label])
  return ids

def oneHotEncodeLabel(labels):
  oneHotClasses = np.zeros([numUniqueLabels])
  for label in labels:
    # Onehot encode each label in the list
    oneHotClasses[label] = 1.
  return oneHotClasses.tolist()

train['LabelIds'] = train['Labels'].apply(mapLabelToId)
train['EncodedLabels'] = train['LabelIds'].apply(oneHotEncodeLabel)

print(train['Labels'][3])
print(train['LabelIds'][3])
print(train['EncodedLabels'][3])

tokenLengths = []
for x in range(train.shape[0]):
    tokenLengths.append(len(train["Tokens"][x]))

maxTokenLength = max(tokenLengths)
print('Max token Length:', maxTokenLength)

# Initialise vocabulary including padding and unknown word tokens
# Is end of sentence needed? '[EOS]'
vocabulary = set(['[PAD]', '[UNK]'])

for tokens in train["Tokens"]:
  vocabulary.update(tokens)

vocabulary = sorted(vocabulary)
print(vocabulary)
print('Unique words:', len(vocabulary))

dictionary = {}
idToDictionary = {}
for idx, word in enumerate(vocabulary):
  dictionary[word] = idx
  idToDictionary[idx] = word

with open('./processed-data/dictionary.json', 'w', encoding='utf-8') as f:
    json.dump(dictionary, f, ensure_ascii=False, indent=4)

# Add dictionary IDs column to dataframe
def addDictionaryIds(tokens):
    ids = []
    for token in tokens:
      if token in dictionary:
        ids.append(dictionary[token])
      else:
        ids.append(dictionary['[UNK]'])
    return ids

train['DictionaryIds'] = train['Tokens'].apply(addDictionaryIds)

def addTokenPadding(tokens):
  padding = maxTokenLength - len(tokens)
  return tokens + padding*[dictionary['[PAD]']]

train['TokensWithPadding'] = train['DictionaryIds'].apply(addTokenPadding)
train.head()

# Setup test data

# Get test CSV dataframe
test = extractDataFrame('data/test.csv')

# Tokenise
test['Tokens'] = test['Caption'].apply(tokenise)
# Apply dictionary IDs
test['DictionaryIds'] = test['Tokens'].apply(addDictionaryIds)
# Add padding
test['TokensWithPadding'] = test['DictionaryIds'].apply(addTokenPadding)

test.head()

train.to_csv('./processed-data/train.csv')
test.to_csv('./processed-data/test.csv')
